+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # Changelog
 
-## 0.2.0 (2023-01-17)
+## 0.3.0 (2023-01-19)
 
 #### New Features
 
+* (desktop): :sparkles: unit test example
 * (desktop): :sparkles: created build
 * :sparkles: delay module
 * :sparkles: mcu minimal build
+#### Others
+
+* :rocket: added github actions